Add nfs-utils
util-linux has the mount command and nfs-utils has mount.nfs. Because of
a security issue (see comment in util-linux expression), util-linux
doesn't look in PATH for the mount helpers. This means that if mount.nfs
is installed in the user profile it will not be found by mount!

A workaround for this is to invoke mount.nfs directly, instead of "mount
-t nfs". But by adding nfs-utils to systemPackages we don't have to,
because mount looks for helpers in the system profile.