Fix apcupsd hooks shell escapes
There is a bug in the NixOS apcupsd module that causes hooks to be
evaluated in the shell at configuration build time.

For my current hooks, it means that HOSTNAME is always set to
'localhost', at build time. Wrong! Fix it locally by doing proper
escaping.

NOTE: I intend to fix the NixOS bug. But since it changes the module
interface in a backwards incompatible way, I think I'll wait until just
before the next release, so that stable + unstable branches have the
same interface for the longest time.