nfs.server: add 'subtree_check' for /nfs export
From "man exports":

As a general guide, a home directory filesystem, which is normally
exported at the root and may see lots of file renames, should be
exported with subtree checking disabled.  A filesystem which is mostly
readonly, and at least doesn't see many file renames (e.g. /usr or /var)
and for which subdirectories may be exported, should probably  be
exported with subtree checks enabled.

Fixes this:

  $ sudo exportfs -a
  exportfs: /etc/exports [1]: Neither 'subtree_check' or
  'no_subtree_check' specified for export "192.168.1.0/24:/nix/".
    Assuming default behaviour ('no_subtree_check').
    NOTE: this default has changed since nfs-utils version 1.0.x