Add 'spotify-ripper' as custom package
Just adding the Nix expression, not installed on any machine. I tested it on
my laptop, for use with Mopidy, but in the end went for the official Spotify
client for me and Maria's big event on 2016-08-06. Reasons: (1) Mopidy would
stop advertising zeroconf service from time to time and print scary errors in
the system log (felt unreliable) and (2) we got Internet after all, so being
offline wasn't such a big deal.