Add ssh-keys.nix (keep ssh keys DRY)
Instead of maintaining the keys as string literals in multiple
locations, add them (once) to ssh-keys.nix and import that file where
needed.

Also split the command= field out of the keys, so that the keys
themselves are "pure" and can be used to give different access rights on
different hosts.