gitolite: use separate service for creating /srv directory
In general, it's not safe to add PermissionsStartOnly = true on a
service not under our control. The module that defines the service may
already use ExecStartPre and expect it runs as an unprivileged user,
creating directories or files for that user.

So instead take the safer approach and use a separate service to create
the /srv directory and tell systemd to run that service before gitolite.

The only downside I can think of is that the dependency information is
not checked at compile time. But I think systemd will at least complain
in the journal if we misspelled the service name.