keil-uvision-c51: new package
This Windows-only program was difficult to package. I couldn't find a
way to force the installer to run without the GUI and without asking
questions. The solution I ended up with is to run the installer manually
and make a tarball of $WINEPREFIX afterwards. See comments in the .nix
file for how this is done.