Add system.stateVersion
On a test machine I was bitten by the fact that postgresql changed
dataDir (and superuser) between NixOS 17.03 and 17.09, causing the hydra
service to fail. Using system.stateVersion would have prevented that.

These days the NixOS installer adds this option to the template
configuration.nix file, so I figure it's a good idea to have it set in
all machine configurations. (machines/hydra.nix already has this
setting, as its config was made from scratch with NixOS 16.09.)