Refactor tmux into a standalone tmux-with-config package
Instead of relying on the global /etc/tmux.conf file, create a tmux
wrapper that sources the tmux.conf file directly from the nix store.
This allows using the package on non-NixOS (well, except from perhaps
missing powerline-fonts, but that's a separate issue). It also speeds up
test iterations, since it's much quicker to rebuild the wrapper than to
rebuild NixOS.

To be able to source a new config in a running session, a separate
tmux-source-conf script is created. This is needed since tmux itself
only reads config files when the server is started. That means the
command to source the config must be available from the (original)
tmux.conf. Using $PATH lookup to find a tmux-source-conf script seemed
like a nice and easy way to solve that problem.

The remaining bits of cfg/tmux/default.nix (adding powerline fonts) is
moved to cfg/base-small.nix, to not cause confusion between cfg/tmux and
pkgs/tmux. (E.g. running fzf and typing 'tmux' gives better results this
way.)