pkgs/vim: switch from VAM to vim packages (built-in)
After cleaning up my impure vim config in $HOME, I noticed that VAM
started re-downloading itself (or vim-pi etc.)! A big part of using Nix
is, for me, is to not fetch stuff at runtime. So switch to the
vim built-in package handling mechanism, since that can be used offline.

vim-addon-nix fails to load in the new setup, perhaps it has a
dependency on VAM? (AFAIK, they have the same author.)

I didn't notice difference in startup time with this new plugin/package
manager.