pkgs/mini-ci: fix creating ref symlinks for annotated tags
When a git hook receives an update for an annotated tag (as opposed to
lightweight tag), the new sha1 reference points to the tag object
itself, not the commit that the tag refers to. That lead mini-ci to
create a ref symlink to a non-existing commit, since the commit really
was a tag object. Fix it by dereferencing annotated tags.

Also add a check that the symlink destination exists, to prevent bugs
like this happening again.