pkgs/git: resolve issue about '[diff "word"]' config sections
Today I learned that a gitconfig section like

  [diff "word"]
  textconv=...

is written as -c diff.word.textconv=... on the command line. Improve the
config generator to support that nested section type of config.

With this in place, there is far less need for having a --config flag in
git. So remove that TODO, I can live without it.