pkgs/mini-ci: handle repos with relative submodule URLs
After copying the repo to a temporary directory (to be able to update
submodules without the side effect of updating files in the original
repo), we must set the remote.origin.url config setting. If not, git
will try to fetch submodules relative to the worktree, but the worktree
is a temporary/random directory where no other git repos exist.