base-medium.nix: set pulseaudio flat-volumes=no
This fixes the annoying behaviour that Kodi (seemingly) sets the master
volume to 100% all the time (or to whatever the kodi volume is set to).
In fact, it is pulseaudio that sets the master volume, to the max of any
application volume. Ref. "man pulse-daemon.conf".

(NixOS 19.03 sets this to "no" by default, like several other distros
have done for quite some time. Upstream pulseaudio still defaults to
"yes", probably because it's easier for newbies.)