options/borg-backup.nix: fix autofs_excludes code
The double quotes around the mountpoint argument gets passed to borg
literally(!). That completely breaks the functionality, causing borg to
exit with error when an autofs mountpoint is unable to mount the
underlying filesystem.

Fix it by using printf %q, which quotes the argument in a way that can
be reused as shell input.