backup-server.nix: use systemd automount
Instead of relying on desktop automount. In preparation for migrating to
a real (headless) server.

Although autofs mountpoints are real mountpoints, we're interested in
the filesystems that get mounted on top. So the "is the disk here" check
must be changed from using "mountpoint" command to checking whether the
device exists in /dev/disk/by-label/.