Add new machine: srv1
My first dedicated server. It's plain desktop hardware, but I won't be
running a graphical desktop environment on it and I'll manage the
machine remotely.

The plan is to migrate services from mini to this machine.

The filesystem stack is like this:

  /boot: mdadm raid 1 (metadata 1.0, superblock at the end) + fat32
  /: mdadm raid 1 (metadata 1.2) + bcache + lvm + ext4

Bcache devices run with a cache set consisting of two NVMe devices in
(sw) raid 1.