Move services from mini to srv1
Services:
* backup-server
* ddclient
* webserver
* cgit
* git-daemon
* gitolite

Service data was moved with rsync when the services were stopped, except
for the backup disk, which was physically moved.

Setup a new local backup job for srv1. The backup job for mini is
changed from a local repo to a ssh:// repo on srv1.

Ownership of /mnt/backup-disk/backups/hosts/mini.local/ was manually set
to backup:backup (was root:root), now that this repo is accessed
remotely from the mini machine through the 'backup' user.