Add ./cfg/nix-path.nix
The plan is to not rely on the channel, because I'm a bit annoyed by how
it first have to be (globally) updated, then you try to build with it,
it might fail and now you have to manually roll back the channel.

The idea is to be able to deploy from arbitrary nixpkgs and have this
module persist the inputs so that any later nixos-rebuild uses the last
known good nixpkgs.