Add popwk - power on/off projector with keyboard
It's a tool / service I wrote in december 2019 (during the holiday) to
make the home cinema projector power on when the (Logitech) keyboard
gets powered on. And power off when the keyboard is powered off, after
having shown a desktop notification for a few seconds.

I implemented some service hardening in january 2020, especially
important since this runs as root. Since then the module has been in use
on the media PC. I've just been sloppy when it comes to adding code back
to the central repository.